requirements.txt
appdirs==1.4.4
asgiref==3.4.1
Cython==0.29.23
distlib==0.3.1
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
greenlet==1.1.0
mysql-connector-python==8.0.25
postgres==3.0.0
protobuf==3.17.2
psycopg2==2.8.6
psycopg2-binary==2.9.1
psycopg2-pool==1.1
pymssql==2.2.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.18
sqlparse==0.4.1
virtualenv==20.4.6
